Update README.md
Stranger Things is a science fiction horror series set in the 1980s that follows a group of kids uncovering mysterious events in their small town of Hawkins, Indiana. The show combines elements of friendship, courage, and supernatural adventure as they face dangers from a secret experiment and a dark parallel world called the Upside Down. Its nostalgic setting, emotional storytelling, and strong characters make it one of the most captivating and unforgettable series of the modern era.